+++ src/base/mod.rs
@@ -1,0 +1,11 @@
+mod getter;
+mod partial_getter;
+mod partial_reversible;
+mod reversible;
+mod setter;
+
+pub use getter::HasGetter;
+pub use partial_getter::HasPartialGetter;
+pub use partial_reversible::HasPartialReversible;
+pub use reversible::HasReversible;
+pub use setter::HasSetter;