+++ src/partial_getter/composed.rs
@@ -1,34 +1,32 @@
-use crate::HasPartialGetter;
-use crate::HasSetter;
-use crate::prism::Prism;
+use crate::{HasPartialGetter, PartialGetter};
 use core::marker::PhantomData;
 
-/// A composed `Prism` type, combining two optics into a single prism.
+/// A composed `PartialGetter` type, combining two optics into a single prism.
 ///
 /// This struct is automatically created by composing two existing optics, and is **not** intended
 /// to be directly constructed outside the crate. Instead, it is generated through composition of
 /// two optics via the corresponding `ComposableXXX` traits, where each optic can be any
-/// valid optic type that results in a `Prism`.
+/// valid optic type that results in a `PartialGetter`.
 ///
-/// A `ComposedPrism` not only combines two optics into a single [`Prism`], but it also inherently
-/// acts as an `Optic`. This behavior arises from the fact that a `Prism` is itself a
-/// more specific form of an optic, and thus any `Prism` composition will also be usable an `Optic`.
+/// A `ComposedPartialGetter` not only combines two optics into a single [`PartialGetter`], but it also inherently
+/// acts as an `Optic`. This behavior arises from the fact that a `PartialGetter` is itself a
+/// more specific form of an optic, and thus any `PartialGetter` composition will also be usable an `Optic`.
 ///
 /// # Construction
 ///
 /// This struct **cannot** be manually constructed by users. Instead, it is created via
 /// composition of two optics using the appropriate `ComposableXXX` trait for each optic type.
-/// The `ComposedPrism` structure is provided internally by the crate after you compose valid optics.
+/// The `ComposedPartialGetter` structure is provided internally by the crate after you compose valid optics.
 ///
 /// # See Also
 ///
-/// - [`Prism`] — the optic type that `ComposedPrism` is based on
+/// - [`PartialGetter`] — the optic type that `ComposedPartialGetter` is based on
 /// - [`Optic`] — the base trait that all optic types implement
 /// - [`crate::composers::ComposableLens`] — a trait for composing [`Lens`] optics another [`Optic`]
-/// - [`crate::composers::ComposablePrism`] — a trait for composing [`Prism`] optics another [`Optic`]
+/// - [`crate::composers::ComposablePartialGetter`] — a trait for composing [`PartialGetter`] optics another [`Optic`]
 /// - [`crate::composers::ComposableIso`] — a trait for composing [`Iso`] optics into another [`Optic`]
 /// - [`crate::composers::ComposableFallibleIso`] — a trait for composing [`FallibleIso`] optics into another [`Optic`]
-pub struct ComposedPrism<O1: Prism<S, I>, O2: Prism<I, A>, E, S, I, A> {
+pub struct ComposedPartialGetter<O1: PartialGetter<S, I>, O2: PartialGetter<I, A>, E, S, I, A> {
     optic1: O1,
     optic2: O2,
     error_fn_1: fn(O1::GetterError) -> E,
@@ -36,10 +34,10 @@ pub struct ComposedPrism<O1: Prism<S, I>, O2: Prism<I, A>, E, S, I, A> {
     _phantom: PhantomData<(S, I, A, E)>,
 }
 
-impl<O1, O2, E, S, I, A> ComposedPrism<O1, O2, E, S, I, A>
+impl<O1, O2, E, S, I, A> ComposedPartialGetter<O1, O2, E, S, I, A>
 where
-    O1: Prism<S, I>,
-    O2: Prism<I, A>,
+    O1: PartialGetter<S, I>,
+    O2: PartialGetter<I, A>,
 {
     pub(crate) fn new(
         optic1: O1,
@@ -47,7 +45,7 @@ where
         error_fn_1: fn(O1::GetterError) -> E,
         error_fn_2: fn(O2::GetterError) -> E,
     ) -> Self {
-        ComposedPrism {
+        ComposedPartialGetter {
             optic1,
             optic2,
             error_fn_1,
@@ -57,10 +55,10 @@ where
     }
 }
 
-impl<O1, O2, E, S, I, A> HasPartialGetter<S, A> for ComposedPrism<O1, O2, E, S, I, A>
+impl<O1, O2, E, S, I, A> HasPartialGetter<S, A> for ComposedPartialGetter<O1, O2, E, S, I, A>
 where
-    O1: Prism<S, I>,
-    O2: Prism<I, A>,
+    O1: PartialGetter<S, I>,
+    O2: PartialGetter<I, A>,
 {
     type GetterError = E;
 
@@ -70,31 +68,18 @@ where
     }
 }
 
-impl<O1, O2, E, S, I, A> HasSetter<S, A> for ComposedPrism<O1, O2, E, S, I, A>
+impl<O1, O2, E, S, I, A> PartialGetter<S, A> for ComposedPartialGetter<O1, O2, E, S, I, A>
 where
-    O1: Prism<S, I>,
-    O2: Prism<I, A>,
-{
-    fn set(&self, source: &mut S, value: A) {
-        if let Ok(mut i) = self.optic1.try_get(source).map_err(self.error_fn_1) {
-            self.optic2.set(&mut i, value);
-            self.optic1.set(source, i);
-        }
-    }
-}
-
-impl<O1, O2, E, S, I, A> Prism<S, A> for ComposedPrism<O1, O2, E, S, I, A>
-where
-    O1: Prism<S, I>,
-    O2: Prism<I, A>,
+    O1: PartialGetter<S, I>,
+    O2: PartialGetter<I, A>,
 {
 }
 
-pub fn new<S, A, I, E, L1: Prism<S, I>, L2: Prism<I, A>>(
+pub fn new<S, A, I, E, L1: PartialGetter<S, I>, L2: PartialGetter<I, A>>(
     l1: L1,
     l2: L2,
     error_fn_1: fn(L1::GetterError) -> E,
     error_fn_2: fn(L2::GetterError) -> E,
-) -> ComposedPrism<L1, L2, E, S, I, A> {
-    ComposedPrism::new(l1, l2, error_fn_1, error_fn_2)
+) -> ComposedPartialGetter<L1, L2, E, S, I, A> {
+    ComposedPartialGetter::new(l1, l2, error_fn_1, error_fn_2)
 }